reorganize select
