+++ packages/select/src/ListMenu/index.ts
@@ -1,0 +1,4 @@
+import ListMenu from './ListMenu';
+
+export default ListMenu;
+export { popoverClassName } from './ListMenu.styles';
