+++ packages/select/src/OptionGroup/index.ts
@@ -1,0 +1,5 @@
+export {
+  InternalOptionGroup,
+  OptionGroup,
+  OptionGroupElement,
+} from './OptionGroup';
