+++ packages/select/src/Select/index.ts
@@ -1,0 +1,7 @@
+export { default as Select } from './Select';
+export {
+  DropdownWidthBasis,
+  type SelectProps,
+  Size,
+  State,
+} from './Select.types';
