+++ packages/select/src/SelectContext.tsx
@@ -2,7 +2,7 @@ import { createContext } from 'react';
 
 import { Theme } from '@leafygreen-ui/lib';
 
-import { Size } from './types';
+import { Size } from './Select/Select.types';
 
 interface SelectData {
   theme: Theme;
